// chaincode/global/enums/Controller.ts
/**
 * Copyright (C) 2020, RTE (http://www.rte-france.com)
 * SPDX-License-Identifier: Apache-2.0
*/


export enum Controller {
  PowerPlanEnergySchedule = 'PowerPlanEnergySchedule',
  EDA = 'EDA',
  EDP = 'EDP',
  MV = 'MV',
  HV = 'HV',
  LogOrder = 'LogOrder',
  Bid = 'Bid',
  OrderBySite = 'OrderBySite',
  Point = 'Point',
  Site = 'Site',
  Activation = 'Activation'
}